+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/fir/firTreeNode.kt
@@ -15,8 +15,7 @@ public fun FirElement.toFirTreeNode(): FirTreeNode {
             override fun visitElement(element: FirElement) {
                 firTree.addChild(element.toFirTreeNode())
             }
-        }
-    )
+        })
     return firTree
 }