+++ kotlin-compiler-test-extensions/main/kotlin/com/javiersc/kotlin/compiler/test/services/AdditionalFilesDirectives.kt
@@ -12,6 +12,5 @@ public object AdditionalFilesDirectives : SimpleDirectivesContainer() {
                     Adds common context annotations
                     See file ./test-data/additional-files/SomeFile.kt
                 """
-                    .trimIndent()
-        )
+                    .trimIndent())
 }